Loading inverted index, and calculating cosine similarity correctly with new method
