+++ ranker/ranker.py
@@ -12,13 +12,14 @@ from mongoengine import *
 from mongoConfig import *
 from cosineSimilarity import *
 import rankerDBConfig
+from loadInvertedIndexToMemory import loadInvertedIndexToMemory
 
 app = Flask(__name__)
 
 port = 8002
 connect(rankerDBConfig.databaseName, host=rankerDBConfig.databaseAddr, port=27017)
 
-inMemoryTFIDF = loadInvertedIndexToMemory()
+inMemoryTFIDF, crawlerReverseMap = loadInvertedIndexToMemory()
 
 @app.route('/', methods=["GET"])
 def index():
@@ -28,7 +29,7 @@ def index():
 def rankQuery(query):
   log('Ranker', 'Received query: '+query)
   queryTerms = stemQuery(query)
-  sortedDocIds=calculateAllCosineSimilarity(queryTerms, inMemoryTFIDF)
+  sortedDocIds=calculateAllCosineSimilarity(queryTerms, inMemoryTFIDF, crawlerReverseMap)
   return helpers.sendPacket(1, 'Successfully retrieved query', {'sortedDocIds':sortedDocIds})
 
 @app.route('/fetchDocuments', methods=['POST'])
