+++ ranker/rankerDBConfig.py
@@ -1,5 +1,5 @@
-databaseName = 'ChefmateDB'
-# databaseName = 'ChefmateDB_Alt'
+# databaseName = 'ChefmateDB'
+databaseName = 'ChefmateDB_Alt'
 
-databaseAddr = '3.21.167.180'
-# databaseAddr = '18.222.251.5'
+# databaseAddr = '3.21.167.180'
+databaseAddr = '18.222.251.5'
